Port the desktop iframe webivew focus improvements to the web
Fixes #83345

The iframe based webviews had a few fixes for focusing. These should not be specific to desktop  and should also be used on web
