+++ src/vs/workbench/contrib/webview/browser/webviewElement.ts
@@ -4,6 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { addDisposableListener } from 'vs/base/browser/dom';
+import { ThrottledDelayer } from 'vs/base/common/async';
 import { streamToBuffer } from 'vs/base/common/buffer';
 import { CancellationToken } from 'vs/base/common/cancellation';
 import { IDisposable } from 'vs/base/common/lifecycle';
@@ -25,8 +26,12 @@ import { Webview, WebviewContentOptions, WebviewExtensionDescription, WebviewOpt
 import { IWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
 
 export class IFrameWebview extends BaseWebview<HTMLIFrameElement> implements Webview {
+
 	private readonly _portMappingManager: WebviewPortMappingManager;
 	private _confirmBeforeClose: string;
+
+	private readonly _focusDelayer = this._register(new ThrottledDelayer(10));
+	private _elementFocusImpl!: (options?: FocusOptions | undefined) => void;
 
 	constructor(
 		id: string,
@@ -96,6 +101,12 @@ export class IFrameWebview extends BaseWebview<HTMLIFrameElement> implements Web
 		element.style.border = 'none';
 		element.style.width = '100%';
 		element.style.height = '100%';
+
+		this._elementFocusImpl = () => element.contentWindow?.focus();
+		element.focus = () => {
+			this.doFocus();
+		};
+
 		return element;
 	}
 
@@ -154,12 +165,6 @@ export class IFrameWebview extends BaseWebview<HTMLIFrameElement> implements Web
 			endpoint: this.externalEndpoint,
 			confirmBeforeClose: this._confirmBeforeClose,
 		};
-	}
-
-	focus(): void {
-		if (this.element) {
-			this._send('focus');
-		}
 	}
 
 	showFind(): void {
@@ -274,6 +279,53 @@ export class IFrameWebview extends BaseWebview<HTMLIFrameElement> implements Web
 			if (e.data.channel === channel) {
 				handler(e.data.data);
 			}
+		});
+	}
+
+	public focus(): void {
+		this.doFocus();
+
+		// Handle focus change programmatically (do not rely on event from <webview>)
+		this.handleFocusChange(true);
+	}
+
+	private doFocus() {
+		if (!this.element) {
+			return;
+		}
+
+		// Clear the existing focus first if not already on the webview.
+		// This is required because the next part where we set the focus is async.
+		if (document.activeElement && document.activeElement instanceof HTMLElement && document.activeElement !== this.element) {
+			// Don't blur if on the webview because this will also happen async and may unset the focus
+			// after the focus trigger fires below.
+			document.activeElement.blur();
+		}
+
+		// Workaround for https://github.com/microsoft/vscode/issues/75209
+		// Electron's webview.focus is async so for a sequence of actions such as:
+		//
+		// 1. Open webview
+		// 1. Show quick pick from command palette
+		//
+		// We end up focusing the webview after showing the quick pick, which causes
+		// the quick pick to instantly dismiss.
+		//
+		// Workaround this by debouncing the focus and making sure we are not focused on an input
+		// when we try to re-focus.
+		this._focusDelayer.trigger(async () => {
+			if (!this.isFocused || !this.element) {
+				return;
+			}
+			if (document.activeElement && document.activeElement?.tagName !== 'BODY') {
+				return;
+			}
+			try {
+				this._elementFocusImpl();
+			} catch {
+				// noop
+			}
+			this._send('focus');
 		});
 	}
 }
